update vite fs.allow comment to reflect backend/ cross-import
