+++ STANDARDS.md
@@ -284,7 +284,7 @@ Sentri supports **SQLite** (default) and **PostgreSQL** (via `DATABASE_URL` env 
 
 - **Repository pattern**: All DB access goes through repository modules in `backend/src/database/repositories/`. Never write raw SQL in route handlers.
 - **Service layer (`backend/src/services/`)**: When a route handler accumulates more than ~15 lines of business logic — especially logic that's reused across multiple handlers (e.g. provenance shapes, multi-table aggregates, mutation invariants) — extract it into a service module. Routes stay thin (HTTP shape + ACL); services own the business rules. See `services/approvalService.js` (AUTO-003b) for the canonical example: `PROVENANCE_CLEAR`, `humanApproval()`, and `computeStats()` are imported by both `routes/tests.js` and the bulk-action handler, so all four approve/restore paths stay byte-identical.
-- **Shared modules (`shared/` at repo root)**: Cross-side constants (event-type literals, enum strings) that both backend and frontend reference belong in `shared/<thing>.js` and are imported relatively from both sides. Vite is already configured to serve from the repo root via `server.fs.allow` (see `frontend/vite.config.js`). See `shared/activityTypes.js` for the canonical example — the `ACTIVITY_TYPES` enum was extracted after a `"test.approve"` vs `"test.approved"` typo regression.
+- **Cross-side constants**: Event-type literals and enum strings referenced by both backend and frontend live under `backend/src/constants/` (not a repo-root `shared/` directory — the backend Docker image only copies `backend/`, so a sibling top-level `shared/` would break the container build). The frontend reaches into `backend/` via `server.fs.allow: ['..']` in `frontend/vite.config.js` and imports relatively (`../../../backend/src/constants/<thing>.js`). See `backend/src/constants/activityTypes.js` for the canonical example — the `ACTIVITY_TYPES` enum was extracted after a `"test.approve"` vs `"test.approved"` typo regression.
 - **JSON columns**: `steps`, `tags`, `results`, `testQueue`, `credentials`, etc. are stored as JSON strings and auto-serialized/deserialized by the repository layer.
 - **Boolean columns**: `isJourneyTest`, `assertionEnhanced`, `isApiTest` are stored as `0`/`1` integers and converted to `true`/`false` by `testRepo`.
 - **ID generation**: Use `idGenerator.js` for domain objects (projects, tests, runs). Use `uuid` only for internal sub-records.
